+++ src/main/kotlin/dev/mizarc/waystonewarps/infrastructure/persistence/migrations/Migration1_AddWarpIconMeta.kt
@@ -1,0 +1,14 @@
+package dev.mizarc.waystonewarps.infrastructure.persistence.migrations
+
+import co.aikar.idb.Database
+
+class Migration1_AddWarpIconMeta : Migration {
+    override val fromVersion: Int = 0
+    override val toVersion: Int = 1
+
+    override fun migrate(db: Database) {
+        runCatching {
+            db.executeUpdate("ALTER TABLE warps ADD COLUMN iconMeta TEXT;")
+        }
+    }
+}